+++ cdl/tests/features/hdf5/h5browser_app.py
@@ -13,6 +13,7 @@ HDF5 Browser Application test
 # guitest: show
 
 from guidata.qthelpers import exec_dialog, qt_app_context
+from qtpy import QtWidgets as QW
 
 from cdl.env import execenv
 from cdl.tests.data import get_test_fnames
@@ -42,6 +43,17 @@ def test_h5browser() -> None:
     """Test HDF5 browser"""
     with qt_app_context():
         dlg = create_h5browser_dialog(get_test_fnames("*.h5")[-2])
+
+        # Test all buttons:
+        dlg.show()
+        for index in range(dlg.button_layout.count()):
+            widget = dlg.button_layout.itemAt(index).widget()
+            if isinstance(widget, QW.QCheckBox):
+                widget.setChecked(True)
+                widget.setChecked(False)
+            elif isinstance(widget, QW.QPushButton):
+                widget.click()
+
         exec_dialog(dlg)