+++ CHANGELOG.md
@@ -17,6 +17,8 @@ for future and past milestones.
   * Added tabs with information below the graphical preview:
     * Group info: path, textual preview, etc.
     * Attributes info: name, value
+  * Added "Show only supported data" check box: when checked, only supported data
+    (signals and images) are shown in the tree view
 * Macro Panel:
   * Macro commands are now numbered, starting from 1, like signals and images
 * Remote control API (`RemoteProxy` and `LocalProxy`):